Pull out AlertTableHeader to a separate component/file
